+++ src/pipelines/clean_pipeline.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import os
-from utils import estimate_missing_aqi
+from src.utils import estimate_missing_aqi
 from sklearn.impute import SimpleImputer
